Created unit tests for us4oem impl validator and rx channel mapping code.

+++ core/devices/TxRxParameters.cpp
@@ -13,8 +13,8 @@ const TxRxParameters TxRxParameters::US4OEM_NOP = TxRxParameters(
     Interval<uint32>(0, 64),
     0, 0);
 
-size_t getNumberOfNoRxNOPs(const TxRxParamsSequence &seq) {
-    size_t res = 0;
+uint16 getNumberOfNoRxNOPs(const TxRxParamsSequence &seq) {
+    uint16 res = 0;
     for(const auto &op : seq) {
         if(!op.isRxNOP()) {
             ++res;
